Catalyst: Add support for exporting / saving results to disk including an entry in the MainMenu that is only selectable in the right controller
This commit showcases how to implement dialogs that can save files to disk.
It required a couple of changes.

1. We needed to be able to encode our `TrainResult` instances as JSON via codable
   (these are the changes in TrainResult and TrainConfiguration)
2. We needed to explicitly tell the Sandbox to allow saving information.
   That is the "File Access: User Selected File" settings in "Signing and Capabilities"
3. We need to write out the file to disk. We write into our sandbox documents folder
4. We use a `UIDocumentPickerViewController` to display a save dialog.

Open Dialogs work accordingly also using `UIDocumentPickerViewController`.

One notable detail is that we also want to call this function from the MainMenu. We added a new entry "Export Result".
This should only be selectable in the `ResultController` (i.e. when we have a result selected).
To do this, we're using the responder chain and the first responder. The menu items will only activate
if the current first responder does have the appropriate method implemented.

For this to work, we also had to force the `ResultController` to be first responder via `canBecomeFirstResponder`
and `becomeFirstResponder`.