Found a better solution around the hack that the ResultController had to be firstResponder
Now, we use `canPerformAction` as well as message forwarding to only
activate the `export` menu entry if we have a ResultController pushed.

This should happen by default based on the first responder entry,
however it seems this doesn't for Nerau (even though it works fine in
my other example apps).