Support for multiple *different* windows. Mostly
So, in practice, what we need to do here, is to create a NSUserActivity
and, in there, store the payload (i.e. as userInfo) that the new window
should receive in order to set up. However, currently, due to bugs in
Catalina, these payloads seem to not always arrive. So in order to
still support something, this feature currently loads a random entry
if the payload is not there.